Pull tsconfig upgrade code to a separate module. Add a bin script.
