+++ bin/ns-upgrade-tsconfig.cmd
@@ -1,0 +1,1 @@
+@node %~dp0\ns-upgrade-tsconfig %*
