+++ package.json
@@ -7,6 +7,9 @@
     "postinstall": "node postinstall.js",
     "preuninstall": "node preuninstall.js"
   },
+  "bin": {
+    "ns-upgrade-tsconfig": "./bin/ns-upgrade-tsconfig"
+  },
   "nativescript": {
     "hooks": [
       {
